[p2p] fix failing connection score

+++ src/p2p/net_node_common.h
@@ -62,7 +62,7 @@ namespace nodetool
     virtual bool unblock_host(const epee::net_utils::network_address &address)=0;
     virtual std::map<std::string, time_t> get_blocked_hosts()=0;
     virtual std::map<epee::net_utils::ipv4_network_subnet, time_t> get_blocked_subnets()=0;
-    virtual bool add_host_fail(const epee::net_utils::network_address &address, unsigned int score = 1)=0;
+    virtual bool add_host_fail(const epee::net_utils::network_address &address, uint64_t score = 1)=0;
     virtual void add_used_stripe_peer(const t_connection_context &context)=0;
     virtual void remove_used_stripe_peer(const t_connection_context &context)=0;
     virtual void clear_used_stripe_peers()=0;
@@ -124,7 +124,7 @@ namespace nodetool
     {
       return std::map<epee::net_utils::ipv4_network_subnet, time_t>();
     }
-    virtual bool add_host_fail(const epee::net_utils::network_address &address, unsigned int score)
+    virtual bool add_host_fail(const epee::net_utils::network_address &address, uint64_t score)
     {
       return true;
     }
